Structured the code: Separate files for main (running the bot), tradingLogic (the trading strat/logic used), technicalIndicators (the calculation of technical indicators). The function tradingLogic(history, symbol) should from now on always be used when selling or buying stocks.
