+++ main.py
@@ -1,8 +1,11 @@
+# THE MAIN FILE THAT RUNS THE BOT
+
 import alpaca_trade_api as alpaca
 import yfinance as yfinance
 import pandas as pandas
 import numpy as numpy
 from alpacaKeys import alpaca_key_id, alpaca_secret_key
+from tradingLogic import tradingLogic
 
 # Alpaca creds and api
 key_id = alpaca_key_id
@@ -18,102 +21,7 @@ alpacaApi = alpaca.REST(
     'v2'
     )
 
-def ema(values, window):
-    # Args: Array of values, number of periods in ema
-    # Return: Array of EMAs of every value after the initial SMA
-
-    emas = []
-
-    valuesSum = 0
-    try:
-        for i in range(window): # Calculate sma
-            valuesSum += values[i]
-    except IndexError:
-        print("Window too large for closing price history.")
-        return
-
-    sma = valuesSum / window
-    smoothing = 2 / (window + 1)
-
-    prevEma = -1
-    for i in range(window, len(values)): # Calculate ema
-
-        if (i == window):        
-            ema = values[i] * smoothing + sma * (1 - smoothing)
-        else:
-            ema = values[i] * smoothing + prevEma * (1 - smoothing)
-
-        emas.append(ema)
-        prevEma = ema
-
-    return emas
-
-def obv(closingPrices, closingVolumes):
-    # Args: Array of closing prices, array of closing volumes
-    # Return: Array of OBV's for each period
-
-    obvs = []
-
-    obv = 0
-    for i in range(len(closingPrices)):
-        if (i == 0):
-            obv = 0
-        elif (closingPrices[i] > closingPrices[i-1]):
-            obv += closingVolumes[i]
-        elif (closingPrices[i] < closingPrices[i-1]):
-            obv -= closingVolumes[i]
-
-        obvs.append(obv)
-
-    return obvs
-
-def rsi(closingPrices, currPrice, window):
-    # Args: Array of previous closing prices, current stock price, number of periods in rsi
-    # Return: RSI of current price
-
-    totUp = 0
-    totDown = 0
-    try:
-        for i in range(1, window + 1): # Calculate starting avgUp/avgDown
-            if (closingPrices[i] > closingPrices[i-1]):
-                totUp += (closingPrices[i] - closingPrices[i-1]) / closingPrices[i-1]
-            elif (closingPrices[i] < closingPrices[i-1]):
-                totDown += (closingPrices[i-1] - closingPrices[i]) / closingPrices[i-1]
-    except IndexError:
-        print("Window too large for closing price history.")
-        return
-
-    avgUp = totUp / window
-    avgDown = totDown / window
-
-    currUp = 0
-    currDown = 0
-    for i in range(window + 1, len(closingPrices)): # Calculate avgUp/avgDown with smoothing
-        if (closingPrices[i] > closingPrices[i-1]):
-            currUp = (closingPrices[i] - closingPrices[i-1]) / closingPrices[i-1]
-        elif (closingPrices[i] < closingPrices[i-1]):
-            currDown = (closingPrices[i-1] - closingPrices[i]) / closingPrices[i-1]
-        
-        avgUp = (avgUp * (window - 1) + currUp) / window
-        avgDown = (avgDown * (window - 1) + currDown) / window
-        currUp = 0
-        currDown = 0
-
-    if (currPrice > closingPrices[-1]): # Calculate current price change
-        currUp = (currPrice - closingPrices[-1]) / closingPrices[-1]
-    elif (currPrice < closingPrices[-1]):
-        currDown = (closingPrices[-1] - currPrice) / closingPrices[-1]
-
-    if (avgDown == 0): # Prevent division by zero
-        rsi = 100
-    else:
-        rsi = 100 - (100 / (1 + ((avgUp * (window - 1) + currUp) / (avgDown * (window - 1) + currDown))))
-
-    return rsi
-
 def scanPortfolio():
-    # CURRENTLY USES MACD 26/12 TRADING STRATEGY
-    
     # Get all current stocks in the portfolio
     # Get history for each stock
     # Decide whether to sell or keep
@@ -140,18 +48,9 @@ def scanPortfolio():
         history = (yfinance.download(symbol, period="3mo")) # 3 month period of data
 
         try:
-            ema26 = ema(history['Close'], 26)
-            ema12 = ema(history['Close'], 12)
-                
-            lenDiff = len(ema12) - len(ema26)
-            macds = []
-            for j in range(len(ema26)):
-                macds.append(ema12[j + lenDiff] - ema26[j])
-            
-            signalLine = ema(macds, 9)
+            action = tradingLogic(history, symbol)
 
-            # Sell if macd crosses below signal line
-            if (macds[-1] < signalLine[-1]):
+            if (action == "SELL"):
                 alpacaApi.submit_order(
                     symbol=symbol,
                     qty=int(portfolioDF.iloc[i]['qty']),
@@ -172,6 +71,6 @@ def scanPortfolio():
     #46.41, 46.22, 45.64, 46.21, 46.25, 45.71, 46.45, 45.78, 45.35, 44.03, 44.18, 44.22, 44.57, 43.42, 42.66]
 #test3 = [10, 10.15, 10.17, 10.13, 10.11, 10.15, 10.20, 10.20, 10.22, 10.21]
 #test4 = [25200, 30000, 25600, 32000, 23000, 40000, 36000, 20500, 23000, 27500]
-#data = yfinance.download("GOOGL", start="2020-01-01", end="2020-01-07")
-#print(data)
+#history = yfinance.download("GOOGL", start="2020-01-01", end="2020-01-07")
+#print(tradingLogic(history, "GOOGL"))
 scanPortfolio()
